pyproject.toml
[project]
name = "pdbq"
version = "0.1.0"
requires-python = ">=3.11"
description = "Natural-language query agent over PeeringDB data"

dependencies = [
    "anthropic>=0.25.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "duckdb>=0.10.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]

[project.scripts]
pdbq = "cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pdbq"]
